face-detection/requirements.txt
# Python dependencies for BlazeFace ONNX face detection
onnxruntime>=1.16.0
Pillow>=9.0.0
numpy>=1.21.0
